Move `TopLevel` to `toplevel.ts`
