+++ src/spec/index.ts
@@ -14,33 +14,10 @@ import {stack} from '../stack';
 import {TitleParams} from '../title';
 import {Transform} from '../transform';
 import {Dict, hash, vals} from '../util';
-import {ConcatLayout, Datasets, GenericCompositionLayout, TopLevelProperties} from './toplevel';
+import {ConcatLayout, GenericCompositionLayout, TopLevel} from './toplevel';
 
-export type TopLevel<S extends BaseSpec> = S &
-  TopLevelProperties & {
-    /**
-     * URL to [JSON schema](http://json-schema.org/) for a Vega-Lite specification. Unless you have a reason to change this, use `https://vega.github.io/schema/vega-lite/v3.json`. Setting the `$schema` property allows automatic validation and autocomplete in editors that support JSON schema.
-     * @format uri
-     */
-    $schema?: string;
-
-    /**
-     * Vega-Lite configuration object.  This property can only be defined at the top-level of a specification.
-     */
-    config?: Config;
-
-    /**
-     * A global data store for named datasets. This is a mapping from names to inline datasets.
-     * This can be an array of objects or primitive values or a string. Arrays of primitive values are ingested as objects with a `data` property.
-     */
-    datasets?: Datasets;
-
-    /**
-     * Optional metadata that will be passed to Vega.
-     * This object is completely ignored by Vega and Vega-Lite and can be used for custom metadata.
-     */
-    usermeta?: object;
-  };
+export {normalizeTopLevelSpec as normalize} from '../normalize';
+export {TopLevel} from './toplevel';
 
 export type BaseSpec = Partial<DataMixins> & {
   /**
@@ -297,8 +274,6 @@ export function isVConcatSpec(spec: BaseSpec): spec is GenericVConcatSpec<any, a
 export function isHConcatSpec(spec: BaseSpec): spec is GenericHConcatSpec<any, any> {
   return spec['hconcat'] !== undefined;
 }
-
-export {normalizeTopLevelSpec as normalize} from '../normalize';
 
 // TODO: add vl.spec.validate & move stuff from vl.validate to here
